requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
psycopg2
python-dotenv
SQLAlchemy
azure-identity
azure-storage-blob
WTForms
email_validator
msgraph-sdk
Flask[async]
aioflask
jwt
json
requests 
msal